+++ pkgs/intl4x/lib/plural_rules.dart
@@ -4,8 +4,8 @@
 
 /// The appropriate locale-dependent plural form.
 ///
-/// The [PluralRules] object returns the [PluralCategory] of the input [num]
-/// to [PluralRules.select]. While English only uses 'one' and 'other', other
+/// The [PluralRules] object selects a localized value based on input [num]
+/// using [PluralRules.select]. While English only uses 'one' and 'other', other
 /// languages have more plural forms. For example, Arabic uses 'zero', 'one',
 /// 'two', 'few', 'many', and 'other'.
 ///
@@ -15,14 +15,13 @@
 ///
 /// void main() {
 ///   final rules = PluralRules(locale: Locale.parse('en-US'));
-///   print(rules.select(3)); // prints 'PluralCategory.other'
+///   print(rules.select(3, one: 'item', other: 'items')); // prints 'items'
 /// }
 /// ```
+///@docImport 'src/plural_rules/plural_rules.dart';
 library;
 
-import 'src/plural_rules/plural_rules.dart' show PluralCategory, PluralRules;
-
 export 'src/locale/locale.dart' show Locale;
-export 'src/plural_rules/plural_rules.dart' show PluralCategory, PluralRules;
+export 'src/plural_rules/plural_rules.dart' show PluralRules;
 export 'src/plural_rules/plural_rules_options.dart'
     show Digits, PluralType, RoundingMode, TrailingZeroDisplay;